add HW08-gc

+++ HW08-gc/homework/src/main/java/ru/calculator/CalcDemo.java
@@ -17,7 +17,7 @@ public class CalcDemo {
     private static final Logger log = LoggerFactory.getLogger(CalcDemo.class);
 
     public static void main(String[] args) {
-        long counter = 500_000_000;
+        int counter = 500_000_000;
         var summator = new Summator();
         long startTime = System.currentTimeMillis();
